Properties change to server database

+++ src/main/java/com/blog/socialshare/service/PostService.java
@@ -4,6 +4,7 @@ import java.util.Date;
 import java.util.List;
 
 import com.blog.socialshare.model.Post;
+import com.blog.socialshare.model.User;
 import com.blog.socialshare.repository.PostRepository;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -42,8 +43,13 @@ public class PostService {
         return postRepository.findById(id).get();
     }
 
-    public void deletePost(Integer id) {
-        postRepository.deleteById(id);
+    public void deletePost(Integer id, User user) {
+        User postUser = postRepository.findById(id).get().getAuthor();
+        if (postUser.getId() == user.getId()) {
+            postRepository.deleteById(id);
+        } else {
+            throw new RuntimeException("You are not the author of this post");
+        }
     }
 
     public void updatePost(Post updatedPost) {
@@ -117,7 +123,8 @@ public class PostService {
     public List<Post> getPostsBySearchAndAuthorIdAndTagId(String search,
             List<Integer> authorId, List<Integer> tagId, Integer start, Integer limit, Date startDate, Date endDate) {
         Pageable pageable = PageRequest.of(start / limit, limit);
-        return postRepository.getPostsBySearchAndAuthorIdAndTagId(search, authorId, tagId, startDate, endDate, pageable).getContent();
+        return postRepository.getPostsBySearchAndAuthorIdAndTagId(search, authorId, tagId, startDate, endDate, pageable)
+                .getContent();
     }
 
     public List<Post> getPostsByTagIdAndSorted(Integer start, Integer limit,
@@ -177,7 +184,8 @@ public class PostService {
             sort = Sort.by(string).descending();
         }
         Pageable pageable = PageRequest.of(start / limit, limit, sort);
-        return postRepository.getPostsBySearchAndAuthorId(searchQuery, authorIds, startDate, endDate, pageable).getContent();
+        return postRepository.getPostsBySearchAndAuthorId(searchQuery, authorIds, startDate, endDate, pageable)
+                .getContent();
     }
 
     public List<Post> getPostsBySearchAndAuthorIdAndTagIdAndSorted(String searchQuery, Integer start, Integer limit,
